# swagger-ui/imdb-movie-openapi3.0-schema.yaml
openapi: 3.0.3
info:
  title: IMDB Movie Discussion API Documentation
  description: |-
    This is a IMDB project API spec based on the OpenAPI 3.0 specification.  All the api specications assume requests will go through Load Balancer, the paths are mentioned based on load balancer path definitions.
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: http://localhost:8005
...
paths:
...
  /movie-service/api/v1/{pk}/reviews/create/:
    post:
      tags:
        - Movie Service
      summary: Create Review for a movie.
      description: Create Review for a movie.
      operationId: reviewCreate
      parameters:
        - name: pk
          in: path
          description: ID of movie to create review
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          description: Information missing or movie was already reviewed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Must be authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              error_code:
                type: string
              error_details:
                type: string
...
    Review:
      type: object
      properties:
        review_user:
          type: number
        rating:
          type: number
        description:
          type: string
        active:
          type: boolean
        created:
          type: string
          format: date-time
        update:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT